feat!: follow the Gherkin standard by not differentiating between step types (#199)
Previously a each type of step definition (`Given`, `When` and `Then`) was given its own category. That means the following steps were considered different:
```cpp
GIVEN("I am different"){}
When("I am different"){}
Then("I am different"){}
```
With this change the above are now considered duplicate steps and they'll result in an ambiguous error. This is considered a breaking change.